Refactor JobMover into separate file, split up functions
Also rename from CompressionJobMover->JobMover, split out functions so
that logic is easier to follow, revise comments. Addressing comments
from Matvey's review, but wanted as separate commit for now.
